refactor(用例脑图):   用例脑图代码结构拆分&mock 表格显示子模块设置
