+++ frontend/src/components/business/ms-minders/locale/en-US.ts
@@ -1,6 +1,10 @@
 export default {
+  'ms.minders.allModule': 'All Modules',
   'ms.minders.precondition': 'Precondition',
   'ms.minders.stepDesc': 'Step Description',
   'ms.minders.stepExpect': 'Expected Result',
   'ms.minders.textDesc': 'Text Description',
+  'ms.minders.caseName': 'Test Case Name',
+  'ms.minders.caseNameNotNull': 'Test Case Name cannot be empty',
+  'ms.minders.commentTotal': '{num} Comments in Total',
 };
