+++ frontend/src/components/pure/ms-list/index.vue
@@ -45,15 +45,15 @@
             </div>
             <slot name="itemRight" :item="item" :index="index"></slot>
           </div>
-          <div
-            v-if="props.mode === 'remote' && index === props.data.length - 1"
-            class="flex h-[32px] items-center justify-center"
-          >
-            <div v-if="noMoreData" class="text-[var(--color-text-4)]">{{ t('ms.timeline.noMoreData') }}</div>
-            <a-spin v-else />
-          </div>
         </div>
       </slot>
+      <div
+        v-if="props.mode === 'remote' && index === props.data.length - 1"
+        class="flex h-[32px] items-center justify-center"
+      >
+        <div v-if="noMoreData" class="text-[var(--color-text-4)]">{{ t('ms.timeline.noMoreData') }}</div>
+        <a-spin v-else />
+      </div>
     </template>
     <template v-if="$slots['empty'] || props.emptyText" #empty>
       <slot name="empty">
